Introduce a pl.actual_size(ref_idx) utility to make masking in pallas easier.
Today, in pallas, if you have a kernel grid where the shape of a blockspec does not line up with an input evenly (aka shape_dim % block_dim != 0) - you will get a shape filled with padding. Users mask this via knowledge of their inputs, but this has two main problems:

1) The mask is always run and computed
2) The kernel has to access data outside of itself, more or less, to build the mask, either via closing over locals or extending the kernel for new inputs.

This CL introduces a util, pl.actual_size, to help with getting the real size of a given ref within a given tile.

Ex:

A reduce sum kernel launched with input x=jnp.ones((9, 8, 128)) and a grid of (2, 1, 1) over blockspec of (8, 8, 128) will have an x_ref of (1, 8, 128) real data and the rest padding in the 2nd tile.

This is the first step in two useful directions:

1) A pl.mask_to_actual_size(ref, val) util that runs conditionally (only when we are on a tile where we expect padding, saving compute)

2) Extending pl.actual_size and pl.mask_to_actual_size to handle jumble/ragged types

PiperOrigin-RevId: 677023684