[Pallas] Add tests for scalar elementwise operations
On TPU, instructions differentiate between vectors and scalars, and the corresponding lowering paths are different. Existing Pallas tests only test vector version of operations, but not the scalar version of them. This PR adds tests for scalar elementwise operations.

The structure of the test is similar to the vector version of the tests above.

PiperOrigin-RevId: 684569107