Add an example for executing serialized function in cpp and refactor the examples dir.
Added api in experimental to serialize and deserialize functions without pickling.

Renamed jax_cpp to cpp_exec to be clear what the examples in the directory are for. The original example there executes a HLO, hence the moving of the example to hlo.

Note that the hlo execution examples seems to be broken as I couldn't find jax/tools/jax_to_hlo.py in the repo.