Remove int4 from jtu.dtypes.all_dtypes.
Why? It's not included in the supported() enumeration on any platforms, so
there is no need to mention it here. I tried fixing this by including it in
supported(), but this led to many errors. It's better to not list it here,
because it might mislead us into thinking it's being tested.