[shape_poly] Fix the handling of jvp(lax.sort)
Previously, `jvp(lax.sort)` used a shape-dependent dtype, for
the types of indices (either `int32` or `int64`, depending on
the size of the dimension). For shape polymorphism, input shapes
can affect other intermediate shapes, but not `dtype`s.

In this case it is easy to just use `int46` independent of
the actual shape.