Rework custom hermetic python instructions.
The focus was shifted from how one should build custom python, as it seems like people don't really have issues with that and the process is fairly standard.

Instead the focus was made on demystifying of what hermetic (custom or not) Python actually is and explaining how a user can customize the build while still keeping it as close to a regular Python workflow as possible.

PiperOrigin-RevId: 700863168