[Mosaic GPU] Improve the implementation of max and exp
Both are very important for FlashAttention and both were poorly mapped to PTX.
For exp, we really do not care about denormals when running in approximate mode,
since they would produce results so close to 1 that it really doesn't matter.
For max, LLVM ended up generating a while bunch of comparisons and selects and
failed to take advantage of the max instructions present in GPUs.

Both of those changes _significantly_ improve the performance of Mosaic attention
kernels for heads smaller than 256 (when the pointwise part dominates the execution
time). In one example I looked at, the utilization jumps from 55% to 64%.

PiperOrigin-RevId: 701042779